Lab 2 - Logically Linked DBs/User_Service/requirements.txt
grpcio
grpcio-tools
psycopg2-binary
PyJWT
requests